backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.3.1
jsonschema==4.20.0
spacy==3.7.2
python-dateutil==2.8.2
